+++ part2/courseinfo/src/components/Content.jsx
@@ -1,0 +1,11 @@
+import { Part } from './Part.jsx'
+
+const Content = ({ part }) => {
+  return (
+    <>
+        {part.map(value => (<Part key={value.id} part={ value }/>))}
+    </>
+  )
+}
+
+export { Content }
