+++ part2/courseinfo/src/components/Header.jsx
@@ -1,0 +1,3 @@
+const Header = ({ course }) => <h1>{course.name}</h1>
+
+export { Header }
