+++ part2/courseinfo/src/components/Part.jsx
@@ -1,0 +1,9 @@
+const Part = ({ part }) => {
+  return (
+    <p>
+        {part.name} {part.exercises}
+    </p>
+  )
+}
+
+export { Part }
