+++ part2/courseinfo/src/components/TotalExercises.jsx
@@ -1,0 +1,10 @@
+const TotalExercises = ({ part }) => {
+  const total = part.reduce((acc, item) => {
+    return (acc += item.exercises)
+  }, 0)
+  return (
+    <p><strong>total of {total} exercises</strong></p>
+  )
+}
+
+export { TotalExercises }
